% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lttb.R
\name{LTTB}
\alias{LTTB}
\title{Downsample a time series using LTTB}
\usage{
LTTB(data, n_bins)
}
\arguments{
\item{data}{Numeric matrix with 2 columns, sorted by the first column}

\item{n_bins}{Number of bins used for downsampling}
}
\value{
matrix with the same columns as \code{data} but at most
  \code{n_bins + 2} rows (fewer if \code{n_bins + 2 > nrow(data)})
}
\description{
Downsample a time series using Steinarsson's
\emph{Largest-Triangle-Three-Buckets} algorithm
}
\details{
For a description of the algorithm, see
\url{https://github.com/sveinn-steinarsson/flot-downsample}
and
Sveinn Steinarsson. 2013.
\emph{Downsampling Time Series for Visual Representation.}
MSc thesis. University of Iceland.
}
\examples{
data(timeseries)
downsampled_timeseries = LTTB(timeseries, n_bins=198)
with(as.data.frame(downsampled_timeseries),
     plot(X, Y, type='l'))

\dontrun{
library(ggplot2)
qplot(X, Y, data=as.data.frame(timeseries), geom='line') +
  geom_line(data=as.data.frame(downsampled_timeseries), colour=I('blue'))
}
}

