% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lttb.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Time series data set for testing}
\format{Matrix of two columns (X and Y) and 5000 rows}
\source{
\url{http://flot.base.is/}
}
\usage{
timeseries
}
\description{
This is one of the data sets used to illustrate the original
implementation of LTTB. It was downloaded from
\url{http://flot.base.is/},
converted from JSON to tab-delimited format,
and read into R with the command
\code{timeseries = as.matrix(read.delim(datafile, as.is=TRUE))}.
}
\keyword{datasets}

